refactor library-related backend
